Add cycles to an instruction #2;
cycles were added per instruction and it is updated when instruction is
executed. It is essential to have synchronization between cpu and ppu to
operate properly.

cycles() function and cycles varibale was added to cpu so that we can
get cycles count from cycles() and update corresponding cycles variable.
also ppu needs to have cycles, we have added cycles variable there and
tick() function is added for both cpu and ppu for synchronization
purpose. speed of cpu is 3x faster than ppu. tick() manage that
difference