Add Render methods
CHR-ROM file and ppu vram is read to extract the rgb value of the color
to be rendered. The extracted vlaue is then stored in an array of struct RGB {x,
y, z} having 256 by 240 dimension. All the extracting process was
carried out with proper guideline from nesdev.org/wiki/PPU.